requirements.txt
eventlet==0.33.0
gevent_socketio==0.3.6
pafy==0.5.5
python_vlc==3.0.12118
